Make WinID a 64 bit integer, this should avoid issues with valid Window handles on windows being interpreted as "no wid set".
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@28795 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libvo/video_out.c
@@ -68,7 +68,7 @@ int vo_refresh_rate=0;
 int vo_keepaspect=1;
 int vo_rootwin=0;
 int vo_border=1;
-int WinID = -1;
+int64_t WinID = -1;
 
 int vo_pts=0; // for hw decoding
 float vo_fps=0;
